Move ReplayLog interface definition to its own file.
