+++ decayedlog.go
@@ -2,7 +2,6 @@ package sphinx
 
 import (
 	"bytes"
-	"crypto/sha256"
 	"encoding/binary"
 	"errors"
 	"fmt"
@@ -20,15 +19,10 @@ const (
 
 	// dbPermissions sets the database permissions to user write-and-readable.
 	dbPermissions = 0600
-
-	// sharedHashSize is the size in bytes of the keys we will be storing
-	// in the DecayedLog. It represents the first 20 bytes of a truncated
-	// sha-256 hash of a secret generated by ECDH.
-	sharedHashSize = 20
 )
 
 var (
-	// sharedHashBucket is a bucket which houses the first sharedHashSize
+	// sharedHashBucket is a bucket which houses the first HashPrefixSize
 	// bytes of a received HTLC's hashed shared secret as the key and the HTLC's
 	// CLTV expiry as the value.
 	sharedHashBucket = []byte("shared-hash")
@@ -38,10 +32,6 @@ var (
 	// that a batch is processed more than once.
 	batchReplayBucket = []byte("batch-replay")
 )
-
-// HashPrefix is a statically size, 20-byte array containing the prefix
-// of a Hash256, and is used to detect duplicate sphinx packets.
-type HashPrefix [sharedHashSize]byte
 
 var (
 	// ErrDecayedLogInit is used to indicate a decayed log failed to create
@@ -53,37 +43,8 @@ var (
 	ErrDecayedLogCorrupted = errors.New("decayed log structure corrupted")
 )
 
-// ReplayLog is an interface that defines a new on-disk data structure that
-// contains a persistent log to enable strong replay protection. The interface
-// is general to allow implementations near-complete autonomy. All of these
-// calls should be safe for concurrent access.
-type ReplayLog interface {
-	// Start starts up the on-disk persistent log. It returns an error if
-	// one occurs.
-	Start() error
-
-	// Stop safely stops the on-disk persistent log.
-	Stop() error
-
-	// Get retrieves an entry from the persistent log given its hash prefix. It
-	// returns the value stored and an error if one occurs. Returns
-	// ErrLogEntryNotFound if hash prefix is not in the log.
-	Get(*HashPrefix) (uint32, error)
-
-	// Put stores an entry into the persistent log given a []byte and an
-	// accompanying purposefully general type. It returns an error if the
-	// provided hash prefix already exists in the log.
-	Put(*HashPrefix, uint32) error
-
-	// PutBatch stores
-	PutBatch(*Batch) (*ReplaySet, error)
-
-	// Delete deletes an entry from the persistent log given its hash prefix.
-	Delete(*HashPrefix) error
-}
-
 // DecayedLog implements the PersistLog interface. It stores the first
-// sharedHashSize bytes of a sha256-hashed shared secret along with a node's
+// HashPrefixSize bytes of a sha256-hashed shared secret along with a node's
 // CLTV value. It is a decaying log meaning there will be a garbage collector
 // to collect entries which are expired according to their stored CLTV value
 // and the current block height. DecayedLog wraps boltdb for simplicity and
@@ -280,20 +241,6 @@ func (d *DecayedLog) gcExpiredHashes(height uint32) (uint32, error) {
 	}
 
 	return numExpiredHashes, nil
-}
-
-// hashSharedSecret Sha-256 hashes the shared secret and returns the first
-// sharedHashSize bytes of the hash.
-func hashSharedSecret(sharedSecret *Hash256) HashPrefix {
-	// Sha256 hash of sharedSecret
-	h := sha256.New()
-	h.Write(sharedSecret[:])
-
-	var sharedHash HashPrefix
-
-	// Copy bytes to sharedHash
-	copy(sharedHash[:], h.Sum(nil))
-	return sharedHash
 }
 
 // Delete removes a <shared secret hash, CLTV> key-pair from the
