+++ replaylog.go
@@ -1,0 +1,60 @@
+package sphinx
+
+import (
+	"crypto/sha256"
+)
+
+const (
+	// HashPrefixSize is the size in bytes of the keys we will be storing
+	// in the DecayedLog. It represents the first 20 bytes of a truncated
+	// sha-256 hash of a secret generated by ECDH.
+	HashPrefixSize = 20
+)
+
+// HashPrefix is a statically size, 20-byte array containing the prefix
+// of a Hash256, and is used to detect duplicate sphinx packets.
+type HashPrefix [HashPrefixSize]byte
+
+// hashSharedSecret Sha-256 hashes the shared secret and returns the first
+// HashPrefixSize bytes of the hash.
+func hashSharedSecret(sharedSecret *Hash256) *HashPrefix {
+	// Sha256 hash of sharedSecret
+	h := sha256.New()
+	h.Write(sharedSecret[:])
+
+	var sharedHash HashPrefix
+
+	// Copy bytes to sharedHash
+	copy(sharedHash[:], h.Sum(nil))
+	return &sharedHash
+}
+
+// ReplayLog is an interface that defines a log of incoming sphinx packets,
+// enabling strong replay protection. The interface is general to allow
+// implementations near-complete autonomy. All methods must be safe for
+// concurrent access.
+type ReplayLog interface {
+	// Start starts up the log. It returns an error if one occurs.
+	Start() error
+
+	// Stop safely stops the log. It returns an error if one occurs.
+	Stop() error
+
+	// Get retrieves an entry from the log given its hash prefix. It returns the
+	// value stored and an error if one occurs. It returns ErrLogEntryNotFound
+	// if the entry is not in the log.
+	Get(*HashPrefix) (uint32, error)
+
+	// Put stores an entry into the log given its hash prefix and an
+	// accompanying purposefully general type. It returns ErrReplayedPacket if
+	// the provided hash prefix already exists in the log.
+	Put(*HashPrefix, uint32) error
+
+	// Delete deletes an entry from the log given its hash prefix.
+	Delete(*HashPrefix) error
+
+	// PutBatch stores a batch of sphinx packets into the log given their hash
+	// prefixes and accompanying values. Returns the set of entries in the batch
+	// that are replays and an error if one occurs.
+	PutBatch(*Batch) (*ReplaySet, error)
+}
